backend/ChatService/requirements.txt
asgiref==3.8.1
Django==4.2.13
django-cors-headers==4.0.0
djangorestframework==3.14
pillow==11.1.0
sqlparse==0.5.0
typing-extensions==4.12.1
channels==3.0.4
channels_redis==4.0.0
django-redis
pyjwt
psycopg2-binary
requests